Rename sample project and relevant packages to example
