+++ README.md
@@ -94,7 +94,7 @@ device/emulator has network access.
 
 To enable debugging output, create and set `BUILDKITE_ANALYTICS_DEBUG_ENABLED` environment variable to `true` on your test environment (CI server or local machine).
 
-For instrumented tests debugging, access the variable using `buildConfigField` and pass it through your `MyTestCollector` class. Refer the [sample project](https://github.com/buildkite/test-collector-android/blob/main/sample/) for implementation.
+For instrumented tests debugging, access the variable using `buildConfigField` and pass it through your `MyTestCollector` class. Refer the [example project](https://github.com/buildkite/test-collector-android/blob/main/example/) for implementation.
 
 ## ⚒ Developing
