+++ docs/architecture/ADR-0000-template.md
