+++ example/proguard-rules.pro
