+++ example/src/main/res/xml/backup_rules.xml
