+++ example/src/test/kotlin/com/buildkite/test/collector/android/example/ExampleUnitTestsSet1.kt
@@ -1,0 +1,25 @@
+package com.buildkite.test.collector.android.example
+
+import org.junit.Assert.fail
+import org.junit.Ignore
+import org.junit.Test
+
+class ExampleUnitTestsSet1 {
+
+    @Test
+    fun testWillPass() {}
+
+    @Test
+    fun testWillFail() {
+        fail("Failing unit test")
+    }
+
+    @Test
+    fun testWillFailWithException() {
+        throw Exception("Failing unit test with custom exception")
+    }
+
+    @Test
+    @Ignore("Skipped unit test")
+    fun testWillBeSkipped() {}
+}
