--- sample/src/androidTest/kotlin/com/buildkite/test/collector/android/sample/MyTestCollector.kt
@@ -1,8 +1,0 @@
-package com.buildkite.test.collector.android.sample
-
-import com.buildkite.test.collector.android.InstrumentedTestCollector
-
-class MyTestCollector : InstrumentedTestCollector(
-    apiToken = BuildConfig.BUILDKITE_ANALYTICS_TOKEN,
-    isDebugEnabled = BuildConfig.BUILDKITE_ANALYTICS_DEBUG_ENABLED
-)
