--- sample/src/main/kotlin/com/buildkite/test/collector/android/sample/MainActivity.kt
@@ -1,50 +1,0 @@
-package com.buildkite.test.collector.android.sample
-
-import android.os.Bundle
-import androidx.activity.ComponentActivity
-import androidx.activity.compose.setContent
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.material.MaterialTheme
-import androidx.compose.material.Surface
-import androidx.compose.material.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.tooling.preview.Preview
-import com.buildkite.test.collector.android.sample.ui.theme.ExampleProjectTheme
-
-class MainActivity : ComponentActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContent {
-            ExampleProjectTheme {
-                Surface(
-                    modifier = Modifier.fillMaxSize(),
-                    color = MaterialTheme.colors.background
-                ) {
-                    Column(
-                        verticalArrangement = Arrangement.Center,
-                        horizontalAlignment = Alignment.CenterHorizontally
-                    ) {
-                        Greeting("Hello")
-                    }
-                }
-            }
-        }
-    }
-}
-
-@Composable
-fun Greeting(message: String) {
-    Text(text = message)
-}
-
-@Preview(showBackground = true)
-@Composable
-fun DefaultPreview() {
-    ExampleProjectTheme {
-        Greeting("Android")
-    }
-}
