--- sample/src/main/kotlin/com/buildkite/test/collector/android/sample/ui/theme/Shape.kt
@@ -1,11 +1,0 @@
-package com.buildkite.test.collector.android.sample.ui.theme
-
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material.Shapes
-import androidx.compose.ui.unit.dp
-
-val Shapes = Shapes(
-    small = RoundedCornerShape(4.dp),
-    medium = RoundedCornerShape(4.dp),
-    large = RoundedCornerShape(0.dp)
-)
