+++ support/scripts/example-unit-tests
@@ -1,0 +1,13 @@
+#!/usr/bin/env bash
+
+set -xeu
+set -o pipefail
+
+function finish() {
+  printf "\n---Finished Running Sample Project Unit Tests---\n"
+}
+
+trap finish EXIT
+
+# Run all unit tests of the example project
+./gradlew clean :example:testDebug
