--- support/scripts/sample-instrumented-tests
@@ -1,13 +1,0 @@
-#!/usr/bin/env bash
-
-set -xeu
-set -o pipefail
-
-function finish() {
-  printf "\n---Finished Running Sample Project Instrumented Tests---\n"
-}
-
-trap finish EXIT
-
-# Run all instrumented tests of the sample project
-./gradlew clean :sample:connectedCheck
