--- support/scripts/sample-unit-tests
@@ -1,13 +1,0 @@
-#!/usr/bin/env bash
-
-set -xeu
-set -o pipefail
-
-function finish() {
-  printf "\n---Finished Running Sample Project Unit Tests---\n"
-}
-
-trap finish EXIT
-
-# Run all unit tests of the sample project
-./gradlew clean :sample:testDebug
